add new class Product Management with methods and modified some classes, class Menu deleted
